Add today meals filter and optimize UI performance
